Release v0.6.1
DLL CHANGES ===============================================================

q2java_Engine.c

    Added:  support for Engine.getPerformanceCounter() and 
            Engine.getPerformanceFrequency() methods.

game.c

    Added:  built-in performance measuring code for Win32 platforms
            (#ifdefed out for Unix), typing "sv time" will display
            DLL timing as well as Java timing
    
JAVA CHANGES ==============================================================

q2java.Engine

    Added:  getPerformanceCounter() which returns some number of elapsed
            clock ticks, and getPerformanceFrequency() which tells you how
            many clock ticks occur per second on your particular platform.
            
            The number of ticks/second is platform and OS dependent - on my 
            machine I get about 1.2 million ticks/second since the DLL can 
            use the Win32 QueryPerformanceCounter() functions.  Unix will
            fall back to the Standard C Library clock() function, which 
            might only get you 1000 ticks/second.
            
q2jgame.Game

    Changed: runFrame() now uses the Engine.getPerformanceCounter() method
             instead of System.currentTimeMillis() in gathering timing info.
             
             svcmd_time() updated to use values from 
             Engine.getPerformanceCounter(), and 
             Engine.getPerformanceFrequency() in determining average 
             serverframe runtimes.
            
baseq2.Door
baseq2.spawn.func_button
baseq2.spawn.func_door
baseq2.spawn.func_door_rotating
baseq2.spawn.func_door_secret
baseq2.spawn.func_plat
baseq2.spawn.func_rotating
baseq2.spawn.func_train
baseq2.spawn.func_water

    Changed: several getSpawnArg() methods updated to expect float values 
             (fixing problems _Quinn pointed out)

+++ classes/baseq2/spawn/func_button.java
@@ -47,12 +47,12 @@ public func_button(String[] spawnArgs) throws GameException
 	if (s != null)
 		fEntity.setModel(s);
 
-	fSpeed = getSpawnArg("speed", 40);
+	fSpeed = getSpawnArg("speed", 40.0F);
 	fAccel = getSpawnArg("accel", fSpeed);
 	fDecel = getSpawnArg("decel", fSpeed);	
-	fWait = getSpawnArg("wait", 4);	
+	fWait = getSpawnArg("wait", 4.0F);	
 	fHealth = fMaxHealth = getSpawnArg("health", 0);
-	int lip = getSpawnArg("lip", 3);
+	float lip = getSpawnArg("lip", 3.0F);
 	
 	// setup door sounds
 	if (getSpawnArg("sounds", 0) != 1)
